refactor: 기물 이동 규칙 관련 패키지명 변경
- direction으로 되어있던 규칙 패키지명을 rule로 변경
- 경로에 관한 규칙을 지닌 인터페이스 및 클래스들을 하위 패키지인 route 패키지로 이동
