+++ src/main/java/janggi/domain/piece/Jolbyeong.java
@@ -1,14 +1,14 @@
 package janggi.domain.piece;
 
-import static janggi.domain.direction.Direction.BACK;
-import static janggi.domain.direction.Direction.FRONT;
-import static janggi.domain.direction.Direction.LEFT;
-import static janggi.domain.direction.Direction.RIGHT;
+import static janggi.domain.rule.route.Direction.BACK;
+import static janggi.domain.rule.route.Direction.FRONT;
+import static janggi.domain.rule.route.Direction.LEFT;
+import static janggi.domain.rule.route.Direction.RIGHT;
 
 import janggi.domain.Location;
 import janggi.domain.Side;
-import janggi.domain.direction.Direction;
-import janggi.domain.direction.Route;
+import janggi.domain.rule.route.Direction;
+import janggi.domain.rule.route.Route;
 
 import java.util.List;
