+++ src/main/java/janggi/domain/rule/route/GungSeongRouteProvider.java
@@ -1,13 +1,13 @@
-package janggi.domain.direction;
+package janggi.domain.rule.route;
 
-import static janggi.domain.direction.Direction.BACK;
-import static janggi.domain.direction.Direction.BACK_LEFT;
-import static janggi.domain.direction.Direction.BACK_RIGHT;
-import static janggi.domain.direction.Direction.FRONT;
-import static janggi.domain.direction.Direction.FRONT_LEFT;
-import static janggi.domain.direction.Direction.FRONT_RIGHT;
-import static janggi.domain.direction.Direction.LEFT;
-import static janggi.domain.direction.Direction.RIGHT;
+import static janggi.domain.rule.route.Direction.BACK;
+import static janggi.domain.rule.route.Direction.BACK_LEFT;
+import static janggi.domain.rule.route.Direction.BACK_RIGHT;
+import static janggi.domain.rule.route.Direction.FRONT;
+import static janggi.domain.rule.route.Direction.FRONT_LEFT;
+import static janggi.domain.rule.route.Direction.FRONT_RIGHT;
+import static janggi.domain.rule.route.Direction.LEFT;
+import static janggi.domain.rule.route.Direction.RIGHT;
 
 import janggi.domain.Location;
 import java.util.List;
