+++ src/test/java/janggi/domain/rule/route/StraightRouteProviderTest.java
@@ -1,4 +1,4 @@
-package janggi.domain.direction;
+package janggi.domain.rule.route;
 
 import janggi.domain.Location;
 import java.util.List;
